Update cwd_card_slider.js
As per Sean Gnau:

This may have been incorrectly suggested otherwise in the past - but it is not okay to use aria-hidden on focusable elements.  We cannot hide button controls from screen readers.  Some screen reader users have vision, and use it to help read content, it is not okay to hide controls away from them.